Ekki hægt að skoða forritið nema maður sé loggaður inn

+++ src/main/java/is/hi/hbv501g/hbv501g/Controllers/WorkoutController.java
@@ -14,6 +14,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import java.util.List;
+import javax.servlet.http.HttpSession;
 
 /******************************************************************************
  *  Nafn    : Hópur 7
@@ -31,56 +32,66 @@ public class WorkoutController {
     private WorkoutService workoutService;
     private ExerciseComboService exerciseComboService;
 
+    private UserServiceImplementation userServiceImplementation;
+
     @Autowired
-    public WorkoutController(WorkoutService workoutService, ExerciseComboService exerciseComboService){
+    public WorkoutController(WorkoutService workoutService, ExerciseComboService exerciseComboService, UserServiceImplementation userServiceImplementation){
         this.workoutService = workoutService;
         this.exerciseComboService = exerciseComboService;
+        this.userServiceImplementation = userServiceImplementation;
     }
 
     @RequestMapping("/workouts")
-    public String homePage(Model model, @Param("keyword") String keyword){
-        // Call a method in a service class
-        List<Workout> allWorkouts = workoutService.listAll(keyword);
-        // Add some data to the model
-        model.addAttribute("workouts",allWorkouts);
-        model.addAttribute("keyword", keyword);
-        return "home";
+    public String homePage(Model model, @Param("keyword") String keyword, HttpSession session){
+        if(userServiceImplementation.userLoggedIn(session)) {
+            // Call a method in a service class
+            List<Workout> allWorkouts = workoutService.listAll(keyword);
+            // Add some data to the model
+            model.addAttribute("workouts", allWorkouts);
+            model.addAttribute("keyword", keyword);
+            return "home";
+        }
+        return "redirect:/error_page1";
     }
     @RequestMapping(value = "/addWorkout",method = RequestMethod.GET)
-    public String addWorkoutForm(Workout workout){
-        return "addWorkout";
+    public String addWorkoutForm(Workout workout, HttpSession session){
+        if(userServiceImplementation.userLoggedIn(session)) {
+            return "addWorkout";
+        }
+        return "redirect:/error_page1";
     }
 
     @RequestMapping(value = "/addWorkout",method = RequestMethod.POST)
-    public String addWorkout(Workout workout, BindingResult result,Model model){
-        if(result.hasErrors()){
-            return "addWorkout";
+    public String addWorkout(Workout workout, BindingResult result,Model model, HttpSession session){
+        if(userServiceImplementation.userLoggedIn(session)) {
+            if (result.hasErrors()) {
+                return "addWorkout";
+            }
+            workoutService.save(workout);
+            return "redirect:/workouts";
         }
-        workoutService.save(workout);
-        return "redirect:/workouts";
+        return "redirect:/error_page1";
     }
-    @RequestMapping(value="/delete/{id}",method = RequestMethod.GET
-    )
-    public String deleteWorkout(@PathVariable("id") long id,  Model model){
-        Workout workoutToDelete = workoutService.findByID(id);
-        workoutService.delete(workoutToDelete);
-        return "redirect:/workouts";
+    @RequestMapping(value="/delete/{id}",method = RequestMethod.GET)
+    public String deleteWorkout(@PathVariable("id") long id,  Model model, HttpSession session){
+        if(userServiceImplementation.userLoggedIn(session)) {
+            Workout workoutToDelete = workoutService.findByID(id);
+            workoutService.delete(workoutToDelete);
+            return "redirect:/workouts";
+        }
+        return "redirect:/error_page1";
     }
     @RequestMapping(value = "/workout/{id}",method = RequestMethod.GET)
-    public String openWorkoutForm(@PathVariable("id") long id,Model model){
-        Workout workoutToOpen = workoutService.findByID(id);
-        List<ExerciseCombo> exerciseCombos = exerciseComboService.findByWorkout(workoutToOpen);
-        //List<ExerciseCombo> exerciseCombosToOpen = exerciseComboService.findAll();
-        model.addAttribute("workout", workoutToOpen);
-        model.addAttribute("ExerciseCombos", exerciseCombos);
+    public String openWorkoutForm(@PathVariable("id") long id,Model model, HttpSession session){
+        if(userServiceImplementation.userLoggedIn(session)) {
+            Workout workoutToOpen = workoutService.findByID(id);
+            List<ExerciseCombo> exerciseCombos = exerciseComboService.findByWorkout(workoutToOpen);
+            //List<ExerciseCombo> exerciseCombosToOpen = exerciseComboService.findAll();
+            model.addAttribute("workout", workoutToOpen);
+            model.addAttribute("ExerciseCombos", exerciseCombos);
 
-        return "workout";
+            return "workout";
+        }
+        return "redirect:/error_page1";
     }
-    /*
-    @RequestMapping(value = "/workout/{id}",method = RequestMethod.POST)
-    public String openWorkout(@Param("id") long id,Model model){
-        model.addAttribute("id", id);
-        return "redirect:/";
-    }
-     */
 }
